[Java] Move code from CatalogTool to ArchiveTool class
